fix: replace deprecated 'bold' from pyglet 2.0.x with 'weight' and resolve module conflict - Replace 'bold' argument with 'weight' in pyglet in pyglet.Label inits - Rename 'types.py' to 'custom_types.py' to avoid standard library   naming conflict - Updated version of pyglet in pyproject.toml to reflect current   "standard" version
